+++ packages/desktop/scripts/windowsSign.js
@@ -1,16 +1,7 @@
-exports.default = async function (configuration) {
-  require('child_process').execSync(
-    `java \
-    -jar jsign/jsign-4.1.jar \
-    --keystore jsign/eToken.cfg \
-    --storepass "${process.env.WINDOWS_TOKEN_PASSWORD}" \
-    --storetype PKCS11 \
-    --tsaurl http://timestamp.digicert.com \
-    --alias "${process.env.WINDOWS_TOKEN_ALIAS}" \
-    "${configuration.path}"
-    `,
-    {
-      stdio: 'inherit',
-    },
-  )
-}
+exports.default = async function(configuration) {
+  if(configuration.path){
+    require("child_process").execSync(
+      `smctl sign --keypair-alias=snkeypair --input "${String(configuration.path)}"`
+    );
+  }
+};