+++ packages/web/src/javascripts/Components/Icon/IconPicker.tsx
@@ -7,6 +7,7 @@ import { getEmojiLength } from './EmojiLength'
 import Icon, { isIconEmoji } from './Icon'
 import { IconNameToSvgMapping } from './IconNameToSvgMapping'
 import { IconPickerType } from './IconPickerType'
+import DecoratedInput from '../Input/DecoratedInput'
 
 type Props = {
   selectedValue: VectorIconNameOrEmoji
@@ -15,6 +16,7 @@ type Props = {
   useIconGrid?: boolean
   iconGridClassName?: string
   className?: string
+  autoFocus?: boolean
 }
 
 const TabButton = forwardRef(
@@ -50,7 +52,15 @@ const TabButton = forwardRef(
   },
 )
 
-const IconPicker = ({ selectedValue, onIconChange, platform, className, useIconGrid, iconGridClassName }: Props) => {
+const IconPicker = ({
+  selectedValue,
+  onIconChange,
+  platform,
+  className,
+  useIconGrid,
+  iconGridClassName,
+  autoFocus,
+}: Props) => {
   const iconKeys = useMemo(() => Object.keys(IconNameToSvgMapping), [])
 
   const iconOptions = useMemo(
@@ -114,13 +124,13 @@ const IconPicker = ({ selectedValue, onIconChange, platform, className, useIconG
   }, [])
 
   return (
-    <div className={`flex h-full flex-grow flex-col overflow-auto ${className}`}>
+    <div className={`flex h-full flex-grow flex-col ${className}`}>
       <div className="flex">
         <TabButton label="Icon" type={'icon'} currentType={currentType} selectTab={selectTab} />
         <TabButton label="Emoji" type={'emoji'} currentType={currentType} selectTab={selectTab} />
         <TabButton label="Reset" type={'reset'} currentType={currentType} selectTab={selectTab} />
       </div>
-      <div className={'mt-2 h-full min-h-0 overflow-auto'}>
+      <div className={classNames('mt-1 h-full min-h-0', currentType === 'icon' && 'overflow-auto')}>
         {currentType === 'icon' &&
           (useIconGrid ? (
             <div
@@ -152,17 +162,14 @@ const IconPicker = ({ selectedValue, onIconChange, platform, className, useIconG
           ))}
         {currentType === 'emoji' && (
           <>
-            <div>
-              <input
-                ref={emojiInputRef}
-                autoComplete="off"
-                autoFocus={emojiInputFocused}
-                className="w-full flex-grow rounded border border-solid border-passive-3 bg-default px-2 py-1 text-base font-bold text-text focus:shadow-none focus:outline-none"
-                type="text"
-                value={emojiInputValue as string}
-                onChange={({ target: input }) => handleEmojiChange((input as HTMLInputElement)?.value)}
-              />
-            </div>
+            <DecoratedInput
+              ref={emojiInputRef}
+              autocomplete={false}
+              autofocus={autoFocus ?? emojiInputFocused}
+              type="text"
+              value={emojiInputValue as string}
+              onChange={(value) => handleEmojiChange(value)}
+            />
             <div className="mt-2 text-sm text-passive-0 lg:text-xs">
               Use your keyboard to enter or paste in an emoji character.
             </div>