+++ packages/api/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.26.72](https://github.com/standardnotes/app/compare/@standardnotes/api@1.26.71...@standardnotes/api@1.26.72) (2023-11-27)
+
+**Note:** Version bump only for package @standardnotes/api
+
+## [1.26.71](https://github.com/standardnotes/app/compare/@standardnotes/api@1.26.70...@standardnotes/api@1.26.71) (2023-11-24)
+
+**Note:** Version bump only for package @standardnotes/api
+
+## [1.26.70](https://github.com/standardnotes/app/compare/@standardnotes/api@1.26.69...@standardnotes/api@1.26.70) (2023-11-24)
+
+**Note:** Version bump only for package @standardnotes/api
+
 ## [1.26.69](https://github.com/standardnotes/app/compare/@standardnotes/api@1.26.68...@standardnotes/api@1.26.69) (2023-10-30)
 
 **Note:** Version bump only for package @standardnotes/api