+++ packages/filepicker/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.28.106](https://github.com/standardnotes/app/compare/@standardnotes/filepicker@1.28.105...@standardnotes/filepicker@1.28.106) (2023-11-27)
+
+**Note:** Version bump only for package @standardnotes/filepicker
+
+## [1.28.105](https://github.com/standardnotes/app/compare/@standardnotes/filepicker@1.28.104...@standardnotes/filepicker@1.28.105) (2023-11-24)
+
+**Note:** Version bump only for package @standardnotes/filepicker
+
+## [1.28.104](https://github.com/standardnotes/app/compare/@standardnotes/filepicker@1.28.103...@standardnotes/filepicker@1.28.104) (2023-11-24)
+
+**Note:** Version bump only for package @standardnotes/filepicker
+
 ## [1.28.103](https://github.com/standardnotes/app/compare/@standardnotes/filepicker@1.28.102...@standardnotes/filepicker@1.28.103) (2023-11-04)
 
 **Note:** Version bump only for package @standardnotes/filepicker