+++ packages/mobile/CHANGELOG.md
@@ -3,6 +3,62 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.56.251](https://github.com/standardnotes/app/compare/@standardnotes/mobile@3.56.250...@standardnotes/mobile@3.56.251) (2023-11-27)
+
+**Note:** Version bump only for package @standardnotes/mobile
+
+## [3.56.250](https://github.com/standardnotes/app/compare/@standardnotes/mobile@3.56.249...@standardnotes/mobile@3.56.250) (2023-11-27)
+
+**Note:** Version bump only for package @standardnotes/mobile
+
+## [3.56.249](https://github.com/standardnotes/app/compare/@standardnotes/mobile@3.56.248...@standardnotes/mobile@3.56.249) (2023-11-27)
+
+**Note:** Version bump only for package @standardnotes/mobile
+
+## [3.56.248](https://github.com/standardnotes/app/compare/@standardnotes/mobile@3.56.247...@standardnotes/mobile@3.56.248) (2023-11-24)
+
+**Note:** Version bump only for package @standardnotes/mobile
+
+## [3.56.247](https://github.com/standardnotes/app/compare/@standardnotes/mobile@3.56.246...@standardnotes/mobile@3.56.247) (2023-11-24)
+
+**Note:** Version bump only for package @standardnotes/mobile
+
+## [3.56.246](https://github.com/standardnotes/app/compare/@standardnotes/mobile@3.56.245...@standardnotes/mobile@3.56.246) (2023-11-24)
+
+**Note:** Version bump only for package @standardnotes/mobile
+
+## [3.56.245](https://github.com/standardnotes/app/compare/@standardnotes/mobile@3.56.244...@standardnotes/mobile@3.56.245) (2023-11-23)
+
+**Note:** Version bump only for package @standardnotes/mobile
+
+## [3.56.244](https://github.com/standardnotes/app/compare/@standardnotes/mobile@3.56.243...@standardnotes/mobile@3.56.244) (2023-11-21)
+
+**Note:** Version bump only for package @standardnotes/mobile
+
+## [3.56.243](https://github.com/standardnotes/app/compare/@standardnotes/mobile@3.56.242...@standardnotes/mobile@3.56.243) (2023-11-21)
+
+**Note:** Version bump only for package @standardnotes/mobile
+
+## [3.56.242](https://github.com/standardnotes/app/compare/@standardnotes/mobile@3.56.241...@standardnotes/mobile@3.56.242) (2023-11-20)
+
+**Note:** Version bump only for package @standardnotes/mobile
+
+## [3.56.241](https://github.com/standardnotes/app/compare/@standardnotes/mobile@3.56.240...@standardnotes/mobile@3.56.241) (2023-11-20)
+
+**Note:** Version bump only for package @standardnotes/mobile
+
+## [3.56.240](https://github.com/standardnotes/app/compare/@standardnotes/mobile@3.56.239...@standardnotes/mobile@3.56.240) (2023-11-20)
+
+**Note:** Version bump only for package @standardnotes/mobile
+
+## [3.56.239](https://github.com/standardnotes/app/compare/@standardnotes/mobile@3.56.238...@standardnotes/mobile@3.56.239) (2023-11-20)
+
+**Note:** Version bump only for package @standardnotes/mobile
+
+## [3.56.238](https://github.com/standardnotes/app/compare/@standardnotes/mobile@3.56.237...@standardnotes/mobile@3.56.238) (2023-11-19)
+
+**Note:** Version bump only for package @standardnotes/mobile
+
 ## [3.56.237](https://github.com/standardnotes/app/compare/@standardnotes/mobile@3.56.236...@standardnotes/mobile@3.56.237) (2023-11-11)
 
 ### Bug Fixes