+++ packages/snjs/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.202.113](https://github.com/standardnotes/app/compare/@standardnotes/snjs@2.202.112...@standardnotes/snjs@2.202.113) (2023-11-27)
+
+**Note:** Version bump only for package @standardnotes/snjs
+
+## [2.202.112](https://github.com/standardnotes/app/compare/@standardnotes/snjs@2.202.111...@standardnotes/snjs@2.202.112) (2023-11-24)
+
+**Note:** Version bump only for package @standardnotes/snjs
+
+## [2.202.111](https://github.com/standardnotes/app/compare/@standardnotes/snjs@2.202.110...@standardnotes/snjs@2.202.111) (2023-11-24)
+
+**Note:** Version bump only for package @standardnotes/snjs
+
+## [2.202.110](https://github.com/standardnotes/app/compare/@standardnotes/snjs@2.202.109...@standardnotes/snjs@2.202.110) (2023-11-23)
+
+**Note:** Version bump only for package @standardnotes/snjs
+
 ## [2.202.109](https://github.com/standardnotes/app/compare/@standardnotes/snjs@2.202.108...@standardnotes/snjs@2.202.109) (2023-11-08)
 
 **Note:** Version bump only for package @standardnotes/snjs